Move android files into platform
