+++ src/openrct2-android/app/src/main/CMakeLists.txt
@@ -415,11 +415,11 @@ endif()
 
 # Phase 2: Add Android native asset management source files
 file(GLOB_RECURSE ANDROID_ASSET_SOURCES
-    "${CMAKE_CURRENT_SOURCE_DIR}/cpp/android_asset_manager.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/cpp/android_asset_manager.h"
-    "${CMAKE_CURRENT_SOURCE_DIR}/cpp/android_jni_bridge.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/cpp/startup_profiler.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/cpp/startup_profiler.h"
+    "${CMAKE_SOURCE_DIR}/src/openrct2/platform/android_asset_manager.cpp"
+    "${CMAKE_SOURCE_DIR}/src/openrct2/platform/android_asset_manager.h"
+    "${CMAKE_SOURCE_DIR}/src/openrct2/platform/android_jni_bridge.cpp"
+    "${CMAKE_SOURCE_DIR}/src/openrct2/platform/startup_profiler.cpp"
+    "${CMAKE_SOURCE_DIR}/src/openrct2/platform/startup_profiler.h"
 )
 
 add_library(openrct2 SHARED ${LIBOPENRCT2_SOURCES} ${ANDROID_ASSET_SOURCES})
