+++ src/openrct2/platform/android_asset_manager.h
