+++ src/openrct2/platform/android_jni_bridge.cpp
