+++ src/openrct2/platform/startup_profiler.h
